requirements.txt
flask
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==0.27.6
langchain
faiss-cpu==1.7.4
